requirements.txt
tensorflow-gpu==1.15.0
atari-py==0.2.9
gym==0.18.0
Pillow
keras==2.6.0
numpy==1.20.2
opencv-python==4.5.1.48